requirements.txt
python-telegram-bot==20.0
Pillow==9.5.0
geopy==2.3.0
piexif==1.1.4
exifread==2.3.2
pillow-heif==0.1.3


# pip install -r requirements.txt
